Begin testing Preparator for synthesizing properties; includes fixes for multiple tests using SparkContext
