+++ build.sbt
@@ -9,3 +9,5 @@ libraryDependencies ++= Seq(
   "org.apache.spark"        %% "spark-core"               % "1.4.1" % "provided",
   "org.apache.spark"        %% "spark-mllib"              % "1.4.1" % "provided",
   "org.scalatest"           %% "scalatest"                % "2.2.1" % "test")
+
+parallelExecution in Test := false
